DNM: Manually retrigger ci with pinned abi3audit and old macosenv
The latest release of abi3audit is incorrectly flagging certain symbols
during the release build. To workaround this in the short term we'll
need to pin abi3audit to avoid this. However for the 1.2.0rc1 release we
are blocked from publishing the package on pypi until this is resolved.
This commit manually updates the pyproject.toml to pin abi3audit and
also adjusts the job config to push from a branch instead of the tag.

Als the macos intel jobs were targeting macos-11 environments but this
went EoL on github actions in June [1] and will not work. This commit
bumps this to macos-12 instead.

[1] https://github.blog/changelog/2024-05-20-actions-upcoming-changes-to-github-hosted-macos-runners/